+++ crates/wdl-analysis/src/handlers/common.rs
@@ -1,7 +1,11 @@
 //! Common utilities for LSP handlers.
 
 mod docs;
+mod namespace;
 mod position;
+mod type_eval;
 
 pub use docs::*;
+pub use namespace::*;
 pub use position::*;
+pub use type_eval::*;