+++ crates/wdl-engine/src/eval.rs
@@ -549,6 +549,13 @@ pub trait EvaluationContext: Send + Sync {
     /// Resolves a type name to a type.
     fn resolve_type_name(&self, name: &str, span: Span) -> Result<Type, Diagnostic>;
 
+    /// Gets the enums map for this evaluation context.
+    ///
+    /// Returns `None` if the context has not loaded enums.
+    fn enums(&self) -> Option<&HashMap<String, Enum>> {
+        None
+    }
+
     /// Gets the base directory for the evaluation.
     ///
     /// The base directory is what paths are relative to.
@@ -635,6 +642,15 @@ impl From<ScopeIndex> for usize {
     fn from(index: ScopeIndex) -> Self {
         index.0
     }
+}
+
+/// Represents enums stored in evaluation contexts.
+#[derive(Debug)]
+pub struct Enum {
+    /// The enum type.
+    pub ty: Type,
+    /// Pre-computed map of variant names to their values.
+    pub variants: HashMap<String, Value>,
 }
 
 /// Represents an evaluation scope in a WDL document.